[Projects]: Updated card, list, and component organisation for projects
