+++ react/src/components/ListProjects.jsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from "react";
-import { Link } from "react-router-dom";
-import Loading from "../Loading";
 import { Octokit } from "octokit";
+import CardProject from "./CardProject";
+import Loading from "../Loading";
 
 const token = import.meta.env.GITHUB_TOKEN;
 const octokit = new Octokit({
@@ -9,59 +9,51 @@ const octokit = new Octokit({
 });
 
 const ListProjects = ({ repo }) => {
-  const [daysSinceLastUpdate, setDaysSinceLastUpdate] = useState(0);
+  const [isLoading, setIsLoading] = useState(true);
+  const [repos, setRepos] = useState([]);
 
   useEffect(() => {
-    const lastUpdated = new Date(repo.updated_at);
-    const now = new Date();
-    const diffTime = Math.abs(now - lastUpdated);
-    const diffDays = Math.ceil(diffTime / (1000 * 60 * 60 * 24));
+    const fetchRepos = async () => {
+      try {
+        setIsLoading(true);
+        const data = await octokit.request(
+          "GET /users/katrine-lombardo/repos",
+          {
+            username: "katrine-lombardo",
+            sort: "updated",
+            direction: "desc",
+            headers: {
+              "X-GitHub-Api-Version": "2022-11-28",
+            },
+          }
+        );
+        setRepos(data.data);
+      } catch (error) {
+        console.error(error.message);
+      } finally {
+        setIsLoading(false);
+      }
+    };
+    fetchRepos();
+  }, []);
 
-    let currentDate = new Date();
-    currentDate.setHours(0, 0, 0, 0);
-
-    if (lastUpdated.getTime() === currentDate.getTime()) {
-      setDaysSinceLastUpdate(-1);
-    } else {
-      setDaysSinceLastUpdate(diffDays);
-    }
-  }, [repo]);
+  if (isLoading) {
+    return <Loading />;
+  }
 
   return (
-    <div className='container'>
-      <ul role='list' className='divide-y divide-gray-100 w-full'>
-        <li className='flex justify-between gap-x-6 py-5'>
-          <div className='flex min-w-0 gap-x-4'>
-            <img
-              className='h-12 w-12 flex-none rounded-full bg-gray-50'
-              src='logos/github_logo.webp'
-              alt=''
-            />
-            <div className='min-w-0 flex-auto'>
-              <Link
-                to={repo.html_url}
-                className='text-sm font-semibold leading-6 text-gray-900 text-left'
-              >
-                <p className='text-sm font-semibold leading-6 text-gray-900 text-left'>
-                  {repo.name}
-                </p>
-              </Link>
-              <p className='mt-1 truncate text-xs leading-5 text-gray-500 text-wrap text-left'>
-                {repo.description}
-              </p>
+    <>
+      <h1>My Projects</h1>
+      <div className='h-[calc(100vh-180px)] w-full justify-center flex'>
+        <div className='flex-wrap rounded-lg border-2 p-6 m-6 overflow-y-auto'>
+          {repos.map((repo) => (
+            <div className='mb-4'>
+              <CardProject key={repo.id} repo={repo} />
             </div>
-          </div>
-          <div className='hidden shrink-0 sm:flex sm:flex-col sm:items-end'>
-            <p className='text-sm leading-6 text-gray-900'>{repo.language}</p>
-            <p className='mt-1 text-xs leading-5 text-gray-500'>
-              {daysSinceLastUpdate === -1
-                ? "Updated today"
-                : `Last updated ${daysSinceLastUpdate} days ago`}
-            </p>
-          </div>
-        </li>
-      </ul>
-    </div>
+          ))}
+        </div>
+      </div>
+    </>
   );
 };
