+++ react/src/pages/Projects.jsx
@@ -10,51 +10,12 @@ const octokit = new Octokit({
 });
 
 const Projects = () => {
-  const [isLoading, setIsLoading] = useState(true);
-  const [repos, setRepos] = useState([]);
-
-  useEffect(() => {
-    const fetchRepos = async () => {
-      try {
-        setIsLoading(true);
-        const data = await octokit.request(
-          "GET /users/katrine-lombardo/repos",
-          {
-            username: "katrine-lombardo",
-            sort: "updated",
-            direction: "desc",
-            headers: {
-              "X-GitHub-Api-Version": "2022-11-28",
-            },
-          }
-        );
-        setRepos(data.data);
-      } catch (error) {
-        console.error(error.message);
-      } finally {
-        setIsLoading(false);
-      }
-    };
-    fetchRepos();
-  }, []);
-
-  if (isLoading) {
-    return <Loading />;
-  }
-
   return (
-    <>
-      <h1>My Projects</h1>
-      <div className='h-[calc(100vh-180px)] w-full justify-center flex'>
-        <div className='flex-wrap rounded-lg border-2 p-6 m-6 overflow-y-auto'>
-          {repos.map((repo) => (
-            <div className='mb-4'>
-              <ListProjects key={repo.id} repo={repo} />
-            </div>
-          ))}
-        </div>
+    <div className='columns-1 p-8'>
+      <div className='flex justify-center'>
+        <ListProjects />
       </div>
-    </>
+    </div>
   );
 };
 export default Projects;
